+++ examples/main/main.tf
@@ -9,7 +9,7 @@ module "sharepoint" {
   domain_fqdn                = "contoso.local"
   time_zone                  = "Romance Standard Time"
   auto_shutdown_time         = "1900"
-  front_end_server_count = 0
+  front_end_servers_count = 0
   add_public_ip_address      = "SharePointVMsOnly"
   rdp_traffic_rule        = "10.20.30.40"
   enable_azure_bastion       = false